requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pandas==2.1.4
numpy==1.26.2
fuzzywuzzy==0.18.0
python-levenshtein==0.23.0
python-dotenv==1.0.0
pytest==7.4.3
requests==2.31.0
streamlit==1.28.1
